# api/openapi.yaml
openapi: 3.0.3
info:
  title: Notification Service API
  description: |
    Event-driven notification system delivering messages over SMS, Email,
    and Push channels. Notifications are accepted via this API, queued in
    RabbitMQ, and delivered asynchronously by workers — creation returning
    201 means "accepted for delivery", not "delivered". Track progress by
    polling the resource or (in a later phase) the WebSocket stream.
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Local docker compose stack

# The API is intentionally unauthenticated in this assessment scope.
security: []

tags:
  - name: notifications
    description: Create and track notifications
  - name: system
    description: Health and diagnostics

paths:
  /api/v1/notifications:
    post:
      tags: [notifications]
      summary: Create a notification
      description: |
        Validates and persists a notification, then queues it for
        asynchronous delivery. When `scheduled_at` is set the notification
        is stored as `scheduled` and enqueued at the scheduled time
        instead. Exactly one recipient per notification.
      operationId: createNotification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationRequest'
      responses:
        '201':
          description: Notification accepted for delivery.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '400':
          description: Malformed JSON or failed validation. Every failing field is listed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Request body exceeds the 1 MiB limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags: [notifications]
      summary: List recent notifications
      description: Newest first. Filtering and cursor pagination arrive in a later release.
      operationId: listNotifications
      parameters:
        - name: limit
          in: query
          required: false
          description: Max results (1-100, default 50).
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Recent notifications.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
        '400':
          description: limit is not a positive integer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/notifications/{id}:
    get:
      tags: [notifications]
      summary: Get a notification by ID
      description: Returns the full notification resource including delivery state.
      operationId: getNotification
      parameters:
        - name: id
          in: path
          required: true
          description: Notification UUID.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The notification.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '400':
          description: The id is not a valid UUID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No notification with this ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  /healthz:
    get:
      tags: [system]
      summary: Liveness probe
      description: Process liveness only; dependency readiness is a separate endpoint.
      operationId: healthz
      responses:
        '200':
          description: Process is alive.

  /api/v1/queues:
    get:
      tags: [system]
      summary: Queue depths
      description: Ready-message counts for work queues, retry tiers, and the DLQ. Mounted only when DASHBOARD_ENABLED.
      operationId: queueDepths
      responses:
        '200':
          description: Per-queue ready counts.
          content:
            application/json:
              schema:
                type: object
                properties:
                  queues:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        ready: { type: integer }
        '500':
          $ref: '#/components/responses/InternalError'

  /api/v1/worker:
    get:
      tags: [system]
      summary: Worker pause state
      description: Mounted only when DASHBOARD_ENABLED.
      operationId: getWorkerState
      responses:
        '200':
          description: Current pause state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerState'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [system]
      summary: Pause or resume the worker
      description: Paused workers cancel their queue subscriptions; messages accumulate as ready. Mounted only when DASHBOARD_ENABLED.
      operationId: setWorkerState
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerState'
      responses:
        '200':
          description: Applied state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerState'
        '400':
          description: Malformed body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  /provider/messages:
    post:
      tags: [system]
      summary: Built-in mock provider intake
      description: |
        Testing stand-in for the external provider; the worker's default
        PROVIDER_URL in docker compose points here. Content containing
        FAILME returns 500 (retryable); REJECTME returns 400 (permanent).
        Mounted only when DASHBOARD_ENABLED.
      operationId: providerReceive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, channel, content]
              properties:
                to: { type: string }
                channel: { type: string }
                content: { type: string }
      responses:
        '202':
          description: Accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  messageId: { type: string }
                  status: { type: string, example: accepted }
                  timestamp: { type: string, format: date-time }
        '400':
          description: Malformed request or simulated permanent rejection.
        '500':
          description: Simulated provider outage.

  /api/v1/provider/messages:
    get:
      tags: [system]
      summary: Messages the mock provider received
      description: Newest first, memory-bounded. Mounted only when DASHBOARD_ENABLED.
      operationId: providerMessages
      responses:
        '200':
          description: Received deliveries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        message_id: { type: string }
                        to: { type: string }
                        channel: { type: string }
                        content: { type: string }
                        received_at: { type: string, format: date-time }

components:
  responses:
    InternalError:
      description: Unexpected server error. The correlation ID from the X-Request-ID response header links the error to server logs.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    Channel:
      type: string
      description: Delivery medium.
      enum: [sms, email, push]

    Priority:
      type: string
      description: Processing priority. Defaults to normal.
      enum: [high, normal, low]

    Status:
      type: string
      description: |
        Notification lifecycle state:
        * `pending` — persisted, not yet handed to the queue
        * `scheduled` — waiting for its future delivery time
        * `queued` — in the message queue awaiting a worker
        * `processing` — a worker is delivering it
        * `retrying` — delivery failed, waiting for the next attempt
        * `sent` — provider accepted the message (terminal)
        * `failed` — delivery exhausted or permanently rejected (terminal)
        * `cancelled` — cancelled before delivery (terminal)
      enum: [pending, scheduled, queued, processing, retrying, sent, failed, cancelled]

    CreateNotificationRequest:
      type: object
      required: [recipient, channel, content]
      properties:
        recipient:
          type: string
          description: |
            Channel-specific address: E.164 phone number for sms
            (e.g. +905551234567), email address for email, device token
            for push.
          example: "+905551234567"
        channel:
          $ref: '#/components/schemas/Channel'
        content:
          type: string
          description: |
            Message body. Limits: sms 160 characters, push 512 bytes,
            email 100 KB.
          example: "Your order has shipped."
        priority:
          $ref: '#/components/schemas/Priority'
        scheduled_at:
          type: string
          format: date-time
          description: Future delivery time. Must be in the future when provided.
        idempotency_key:
          type: string
          description: |
            Client-supplied key preventing duplicate creation on retries.
            Reserved — enforced in an upcoming release.

    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        batch_id:
          type: string
          format: uuid
          description: Present when created via a batch request.
        recipient:
          type: string
        channel:
          $ref: '#/components/schemas/Channel'
        content:
          type: string
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/Status'
        scheduled_at:
          type: string
          format: date-time
        attempts:
          type: integer
          description: Delivery attempts made so far.
        last_error:
          type: string
          description: Most recent delivery error, if any.
        provider_message_id:
          type: string
          description: Provider-issued message ID once sent.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        sent_at:
          type: string
          format: date-time

    WorkerState:
      type: object
      required: [paused]
      properties:
        paused:
          type: boolean

    FieldError:
      type: object
      properties:
        field:
          type: string
          example: recipient
        message:
          type: string
          example: must be an E.164 phone number, e.g. +905551234567

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: validation failed
        fields:
          type: array
          description: Present on validation failures; one entry per failing field.
          items:
            $ref: '#/components/schemas/FieldError'